added coordinates to cities

+++ src/main/java/edu/goldenhammer/database/data_types/IDatabaseCity.java
@@ -6,6 +6,6 @@ package edu.goldenhammer.database.data_types;
 public interface IDatabaseCity {
     String getID();
     String getName();
-    double getPointX();
-    double getPointY();
+    int getPointX();
+    int getPointY();
 }
